Move DatatablesParam to reporting code
